+++ package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@artsy/express-reloadable": "^1.1.0",
     "@artsy/stitch": "^1.2.1",
-    "@artsy/reaction-force": "0.13.3",
+    "@artsy/reaction-force": "0.14.0",
     "accounting": "^0.4.1",
     "analytics-node": "^2.4.0",
     "artsy-backbone-mixins": "artsy/artsy-backbone-mixins",
